requirements.txt
python-binance
numpy
backtrader